pyproject.toml
[project]
name = "pge-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=1.38.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
